Rename routes.py to router.py
