+++ weather_cases/exceptions.py
